# Conflicts:
#	src/win32/i_system.cpp
#	tools/re2c/re2c.vcproj

+++ src/win32/i_system.cpp
@@ -1424,8 +1424,8 @@ static HCURSOR CreateBitmapCursor(int xhot, int yhot, HBITMAP and_mask, HBITMAP 
 	ICONINFO iconinfo =
 	{
 		FALSE,		// fIcon
-		xhot,		// xHotspot
-		yhot,		// yHotspot
+		(DWORD)xhot,	// xHotspot
+		(DWORD)yhot,	// yHotspot
 		and_mask,	// hbmMask
 		color_mask	// hbmColor
 	};
